feat: render cohort2 data (#22)
* feat: render cohort2 data

This PR is continous work on consolidating the available data/metrics from PLDG cohort2. For now, it includes a lot of work around cleaning up the codebase so it void of unused dependencies, modules, etc, and ensuring a uniform code style/standard with eslint/prettier

Here's an outline of what has been done:

- updated the active contributors logic to use the GitHub username row from the CSV file instaed of the name row which is pretty much empty
- added missing utility types and interfaces
- setup eslint, prettier and a GitHub action to run this test in a CI environment

* remove redundant console.logs which makes debuggin tedious by cluttering the web console

* fix cache entry

* fix(ui): render cohort2 data successfully

- fixed the date mismatch issue for top performers when parsing the CSV data. Our current CSV has teh date format like so: 'Week 3: January 27, 2025 (Monday) - February 2, 2025 (Sunday)'. The former function went on to match the wrong date.

The edge-case about this now is for us to ensure that we stick to this date format for future cohorts. If not, we may run into the same issue.

- Ensured that a redundant variable, csvDataFiltered, with a dependency on the cohort ID was removed as it doesn't really do anything, because the the state update is managed elsewhere. This change also led to a refactor of the weekly change computation

* run CI on cohort-2 branch also