chore: remove scripts module
The WES (Weekly Engagement Survey) are uploaded to a MongoDB database for now, which can be queried.

However, this isn't really a scalable approach, as we'd have to manually upload data. To solve this, we can pull the data directly from Airtable and convert to a CSV file which most of our functions already parses.